# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My docs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/arm/projects/distribution": {
            "patch": {
                "tags": [
                    "ARM Distribution"
                ],
                "summary": "Обновление распределения студентов по проектам",
                "description": "Позволяет изменить количество мест в проектах и выполнить перераспределение студентов",
                "operationId": "updateDistribution",
                "requestBody": {
                    "description": "Структура с обновленным количеством мест по проектам",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "required": [
                                        "institute_id",
                                        "departments"
                                    ],
                                    "properties": {
                                        "institute_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "departments": {
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "department_id",
                                                    "projects"
                                                ],
                                                "properties": {
                                                    "department_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "projects": {
                                                        "type": "array",
                                                        "items": {
                                                            "required": [
                                                                "project_id",
                                                                "places"
                                                            ],
                                                            "properties": {
                                                                "project_id": {
                                                                    "type": "integer",
                                                                    "example": 101
                                                                },
                                                                "places": {
                                                                    "type": "integer",
                                                                    "example": 4
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление распределения и возврат новой структуры",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "institute_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "institute_name": {
                                                "type": "string",
                                                "example": "Институт Name"
                                            },
                                            "departments": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "department_id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "department_name": {
                                                            "type": "string",
                                                            "example": "Кафедра Name"
                                                        },
                                                        "projects": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "project_id": {
                                                                        "type": "integer",
                                                                        "example": 123
                                                                    },
                                                                    "title": {
                                                                        "type": "string",
                                                                        "example": "Разработка веб-приложения"
                                                                    },
                                                                    "places": {
                                                                        "type": "integer",
                                                                        "example": 4
                                                                    },
                                                                    "candidates_count": {
                                                                        "type": "integer",
                                                                        "example": 3
                                                                    },
                                                                    "specialities": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "object"
                                                                        }
                                                                    },
                                                                    "candidates": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "object"
                                                                        }
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}